+++ CHANGELOG.md
@@ -22,16 +22,19 @@ The version is represented by three digits: a.b.c.
 
 ## Unreleased
 
+FEATURE:
+- `generators`: add `zaks` generator
+
 MAINTENANCE:
-- added mypy
-- change API private methods in module `symmetria.generators`
+- `.pre-commit-config.yaml`: add `mypy`
+- `symmetria.generators`: change API private methods
 
 ## \[0.3.2\] - 2025-02-08
 
 MAINTENANCE:
 - deprecated symmetria.generator: This function will be deprecated in a future version. Use `permutation_generator` instead
 - deprecated symmetria.generators.random.api.random: This function will be deprecated in a future version. Use `random_permutation` instead
-- Makefile: added `ruff` command.
+- `Makefile`: added `ruff` command.
 
 ## \[0.3.1\] - 2024-12-16